chore: move scripts to tutoraspects
